Fixed conversion warning

+++ Group Project - Space Shooter/Project/Player.cpp
@@ -23,7 +23,7 @@ Player::Player(Vector2 v2Position) : Actor(v2Position)
 
 	//Create a varible for the collider, by getting pointer the texture manager, and return the player texture, and ask for its width 
 	//and divide it by 2. Then set that as the extend in the Box Collider
-	Vector2 v2Extend = Vector2((m_pTexture->GetWidth()) / 2, (m_pTexture->GetHeight()) / 2);
+	Vector2 v2Extend = Vector2((float)((m_pTexture->GetWidth()) / 2), (float)((m_pTexture->GetHeight()) / 2));
 	m_pCollider = new BoxCollider(v2Extend);
 	CollisionManager* pCollisionManager;
 	pCollisionManager->GetInstance()->AddObject(this);
@@ -108,7 +108,7 @@ void Player::Update(float fDeltaTime)
 void Player::OnCollision(GameObject* pOtherObject)
 {
 	
-	if (pOtherObject->GetType() == GameObjectType::ROCK) //If colliding with a Rock
+ 	if (pOtherObject->GetType() == GameObjectType::ROCK) //If colliding with a Rock
 	{
 		//Seperate from object
 		SetPosition(m_v2PreviousPosition);
